Rename v3 files
