+++ packages/datetime/src/blueprint-datetime.scss
@@ -2,10 +2,10 @@
 // Licensed under the Apache License, Version 2.0.
 
 @import "common/react-day-picker-overrides";
-@import "components/date-input3/dateinput";
-@import "components/date-picker3/datepicker";
-@import "components/date-picker3/date-picker3";
-@import "components/date-picker3/date-picker3-caption";
-@import "components/date-range-picker3/daterangepicker";
-@import "components/date-range-picker3/date-range-picker3";
+@import "components/date-input/dateinput";
+@import "components/date-picker/datepicker";
+@import "components/date-picker/date-picker";
+@import "components/date-picker/date-picker-caption";
+@import "components/date-range-picker/daterangepicker";
+@import "components/date-range-picker/date-range-picker";
 @import "components/time-picker/timepicker";
