+++ packages/datetime/src/components/date-picker/_date-picker-caption.scss
