+++ packages/datetime/src/components/date-picker/_datepicker.scss
