+++ packages/datetime/src/components/date-range-input/dateRangeInputProps.ts
@@ -5,7 +5,9 @@
 import type { InputGroupProps, Props } from "@blueprintjs/core";
 
 import type { DateFormatProps, DatePickerBaseProps, DateRange } from "../../common";
+import type { DateFnsLocaleProps } from "../../common/dateFnsLocaleProps";
 import type { DatetimePopoverProps } from "../../common/datetimePopoverProps";
+import type { ReactDayPickerRangeProps } from "../../common/reactDayPickerProps";
 import type { DateRangeShortcut } from "../shortcuts/shortcuts";
 
 export interface DateRangeInputProps extends DatePickerBaseProps, DateFormatProps, DatetimePopoverProps, Props {
@@ -124,3 +126,55 @@ export interface DateRangeInputProps extends DatePickerBaseProps, DateFormatProp
      */
     value?: DateRange;
 }
+
+/**
+ * Props shared between DateRangeInput v1 and v
+ *
+ * Note that we exclude formatDate and parseDate so that we can make those optional in DateInput3 and provide a default
+ * implementation for those functions using date-fns.
+ */
+type DateRangeInputSharedProps = Omit<
+    DateRangeInputProps,
+    "dayPickerProps" | "formatDate" | "locale" | "localeUtils" | "modifiers" | "parseDate"
+>;
+
+export interface DateRangeInput3Props
+    extends DateRangeInputSharedProps,
+        ReactDayPickerRangeProps,
+        DateFnsLocaleProps,
+        Partial<Omit<DateFormatProps, "locale">> {
+    /**
+     * [date-fns format](https://date-fns.org/docs/format) string used to format & parse date strings.
+     *
+     * Mutually exclusive with the `formatDate` and `parseDate` props.
+     *
+     * See date-fns [format](https://date-fns.org/docs/format).
+     */
+    dateFnsFormat?: string;
+}
+
+export type DateRangeInput3DefaultProps = Required<
+    Pick<
+        DateRangeInput3Props,
+        | "allowSingleDayRange"
+        | "closeOnSelection"
+        | "contiguousCalendarMonths"
+        | "dayPickerProps"
+        | "disabled"
+        | "endInputProps"
+        | "invalidDateMessage"
+        | "locale"
+        | "maxDate"
+        | "minDate"
+        | "outOfRangeMessage"
+        | "overlappingDatesMessage"
+        | "popoverProps"
+        | "selectAllOnFocus"
+        | "shortcuts"
+        | "singleMonthOnly"
+        | "startInputProps"
+    >
+>;
+
+export type DateRangeInput3PropsWithDefaults = Omit<DateRangeInput3Props, keyof DateRangeInput3DefaultProps> &
+    DateRangeInput3DefaultProps;
