+++ packages/datetime/src/components/date-range-picker/_date-range-picker.scss
