+++ packages/datetime/src/components/date-range-picker/_daterangepicker.scss
